+++ src/typeguard/__init__.py
@@ -202,8 +202,11 @@ def resolve_forwardref(maybe_ref, memo: _TypeCheckMemo):
 
 
 def get_type_name(type_):
-    # typing.* types don't have a __name__ on Python 3.7+
-    return getattr(type_, '__name__', None) or getattr(type_, '_name', None) or str(type_)
+    name = getattr(type_, '__name__', None) or getattr(type_, '_name', None)
+    if name is None or name == 'Literal':
+        return str(type_)
+    else:
+        return name
 
 
 def find_function(frame) -> Optional[Callable]: